[FE-127] Replies to comments.
