--- src/components/Buttons/Buttons.module.css
@@ -1,3 +1,0 @@
-.item {
-  display: inline-block;
-}
