+++ src/components/UserActions/Redeem/RedeemAll/RedeemAll.jsx
@@ -2,25 +2,23 @@ import React, { useState, useEffect } from 'react';
 import { useSubstrate } from '../../../../substrate-lib';
 import { UNDERLYING_ASSETS_TYPES } from '../../../../util/constants';
 
-import { Form, Input, Dropdown, Button } from 'semantic-ui-react';
+import { Form, Dropdown, Button } from 'semantic-ui-react';
 import Loading from '../../../../util/Loading';
-import classes from './RedeemUnderlyingAsset.module.css';
+import classes from './RedeemAll.module.css';
 
-function RedeemUnderlyingAsset({ account, onChange, userState }) {
+function RedeemAll({ account, onChange, userState }) {
 	const { api, keyring } = useSubstrate();
-	const [amount, setAmount] = useState(0);
 	const [asset, setAsset] = useState('');
 	const [loading, setLoading] = useState(false);
 	const [isInvalid, setInvalid] = useState(true);
 
 	useEffect(() => {
-		setInvalid(!(asset && amount && account));
-	}, [setInvalid, account, amount, asset]);
+		setInvalid(!(asset && account));
+	}, [setInvalid, account, asset]);
 
 	const setInitialStates = () => {
-		setAmount(0);
 		setAsset('');
-		setInvalid(!(asset && amount && account));
+		setInvalid(!(asset && account));
 	};
 
 	const assets = UNDERLYING_ASSETS_TYPES.map((currency) => ({
@@ -29,19 +27,15 @@ function RedeemUnderlyingAsset({ account, onChange, userState }) {
 		value: currency,
 	}));
 
-	const onChangeAmount = (e) => {
-		setAmount(e.target.value * 10 ** 18);
-	};
-
 	const onChangeAsset = (e) => {
 		setAsset(e.target.innerText);
 	};
 
-	const redeemUnderlyingAsset = async () => {
+	const sendRedeemAll = async () => {
 		setLoading(true);
 		const currentUser = keyring.getPair(account);
 		await api.tx.minterestProtocol
-			.redeemUnderlying(asset, amount.toString())
+			.redeem(asset)
 			.signAndSend(currentUser, ({ events = [], status }) => {
 				if (status.isFinalized) {
 					setLoading(false);
@@ -73,31 +67,24 @@ function RedeemUnderlyingAsset({ account, onChange, userState }) {
 	}
 
 	return (
-		<div className={classes.redeem}>
-			<Form>
-				<Input
-					type='text'
-					placeholder='Enter the amount'
-					onChange={onChangeAmount}
-				/>
-				<Dropdown
-					compact
-					placeholder='Asset'
-					search
-					selection
-					options={assets}
-					onChange={onChangeAsset}
-				/>
-				<Button
-					color={account ? 'green' : 'red'}
-					onClick={redeemUnderlyingAsset}
-					disabled={isInvalid}
-				>
-					Redeem Underlying Asset
-				</Button>
-			</Form>
-		</div>
+		<Form className={classes.redeem}>
+			<Dropdown
+				compact
+				placeholder='Asset'
+				search
+				selection
+				options={assets}
+				onChange={onChangeAsset}
+			/>
+			<Button
+				color={account ? 'green' : 'red'}
+				onClick={sendRedeemAll}
+				disabled={isInvalid}
+			>
+				Redeem All Asset
+			</Button>
+		</Form>
 	);
 }
 
-export default RedeemUnderlyingAsset;
+export default RedeemAll;
