+++ src/components/UserActions/Redeem/RedeemWrappedToken/RedeemWrappedToken.jsx
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from 'react';
-import { useSubstrate } from '../../../substrate-lib';
-import { UNDERLYING_ASSETS_TYPES } from '../../../util/constants';
-import Loading from '../../../util/Loading';
+import { useSubstrate } from '../../../../substrate-lib';
+import { WRAP_TOKEN_TYPES } from '../../../../util/constants';
 
 import { Form, Input, Dropdown, Button } from 'semantic-ui-react';
+import Loading from '../../../../util/Loading';
 
-import classes from './Borrow.module.css';
+import classes from './RedeemWrappedToken.module.css';
 
-function Borrow({ account, onChange, userState }) {
+function RedeemWrappedToken({ account, onChange, userState }) {
 	const { api, keyring } = useSubstrate();
 	const [amount, setAmount] = useState(0);
 	const [asset, setAsset] = useState('');
@@ -24,7 +24,7 @@ function Borrow({ account, onChange, userState }) {
 		setInvalid(!(asset && amount && account));
 	};
 
-	const assets = UNDERLYING_ASSETS_TYPES.map((currency) => ({
+	const assets = WRAP_TOKEN_TYPES.map((currency) => ({
 		key: currency,
 		text: currency,
 		value: currency,
@@ -38,11 +38,11 @@ function Borrow({ account, onChange, userState }) {
 		setAsset(e.target.innerText);
 	};
 
-	const sendDeposit = async () => {
+	const redeemWrappedToken = async () => {
 		setLoading(true);
 		const currentUser = keyring.getPair(account);
 		await api.tx.minterestProtocol
-			.borrow(asset, amount.toString())
+			.redeemWrapped(asset, amount.toString())
 			.signAndSend(currentUser, ({ events = [], status }) => {
 				if (status.isFinalized) {
 					setLoading(false);
@@ -74,31 +74,29 @@ function Borrow({ account, onChange, userState }) {
 	}
 
 	return (
-		<div className={classes.borrow}>
-			<Form>
-				<Input
-					type='text'
-					placeholder='Enter the amount'
-					onChange={onChangeAmount}
-				/>
-				<Dropdown
-					compact
-					placeholder='Asset'
-					search
-					selection
-					options={assets}
-					onChange={onChangeAsset}
-				/>
-				<Button
-					color={account ? 'green' : 'red'}
-					onClick={sendDeposit}
-					disabled={isInvalid}
-				>
-					Borrow
-				</Button>
-			</Form>
-		</div>
+		<Form className={classes.redeem}>
+			<Input
+				type='text'
+				placeholder='Enter the amount'
+				onChange={onChangeAmount}
+			/>
+			<Dropdown
+				compact
+				placeholder='Asset'
+				search
+				selection
+				options={assets}
+				onChange={onChangeAsset}
+			/>
+			<Button
+				color={account ? 'green' : 'red'}
+				onClick={redeemWrappedToken}
+				disabled={isInvalid}
+			>
+				Redeem Wrapped Token
+			</Button>
+		</Form>
 	);
 }
 
-export default Borrow;
+export default RedeemWrappedToken;
